Refactor video modal zap flow into dedicated controller
